+++ dawheen_backend/src/main/java/com/study/dawheen/organization/repository/impl/OrganRepositoryCustomImpl.java
@@ -23,7 +23,7 @@ public class OrganRepositoryCustomImpl implements OrganRepositoryCustom {
     }
 
     @Override
-    public Optional<List<Organization>> findOrganizationsWithinRadius(double latitude, double longitude, int radius) {
+    public Optional<List<Organization>> findOrganizationsWithinRadius(double latitude, double longitude, double radius) {
 
         QOrganization organization = QOrganization.organization;
         QCoordinate coordinate = QCoordinate.coordinate;
